+++ src/felupe/constitution/strain/_material_strain.py
@@ -18,18 +18,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 
-from ...math import (
-    cdya_ik,
-    cof,
-    dot,
-    identity,
-    inv,
-    ravel,
-    reshape,
-    svd,
-    sym,
-    transpose,
-)
+from ...math import cdya_ik, dot, identity, ravel, reshape, sym, transpose
 from .._base import ConstitutiveMaterial
 
 
@@ -53,7 +42,10 @@ class MaterialStrain(ConstitutiveMaterial):
     framework : str, optional
         The framework to be used for the stress and strain formulations. "small-strain"
         and "total-lagrange" are supported. Default is "small-strain".
-
+    symmetry : bool, optional
+        Take the symmetric part of the returned stress and the minor symmetric-parts of
+        the algorithmic consistent elasticity tensor. Default is True. May enhance
+        performance if the material returns symmetric tensors.
 
     Examples
     --------
@@ -63,7 +55,7 @@ class MaterialStrain(ConstitutiveMaterial):
         :context: close-figs
 
         import felupe as fem
-        from felupe.math import identity, cdya, dya, trace
+        from felupe.math import identity, cdya_ik, dya, trace
 
         def linear_elastic(dε, εn, σn, ζn, λ, μ, **kwargs):
             '''3D linear-elastic material formulation.
@@ -90,7 +82,7 @@ class MaterialStrain(ConstitutiveMaterial):
 
             # update stress and evaluate elasticity tensor
             σ = σn + dσ
-            dσdε = 2 * μ * cdya(I, I) + λ * dya(I, I)
+            dσdε = 2 * μ * cdya_ik(I, I) + λ * dya(I, I)
 
             # update state variables (not used here)
             ζ = ζn
@@ -141,6 +133,7 @@ class MaterialStrain(ConstitutiveMaterial):
         dim=3,
         statevars=(0,),
         framework="small-strain",
+        symmetry=True,
         **kwargs,
     ):
         self.material = material
@@ -148,7 +141,9 @@ class MaterialStrain(ConstitutiveMaterial):
         self.statevars_size = [np.prod(shape) for shape in statevars]
         self.statevars_offsets = np.cumsum(self.statevars_size)
         self.nstatevars = sum(self.statevars_size)
+
         self.framework = framework
+        self.symmetry = symmetry
 
         self.kwargs = {**kwargs, "tangent": None}
 
@@ -218,8 +213,10 @@ class MaterialStrain(ConstitutiveMaterial):
             axis=0,
         )
 
-        if self.framework == "total-lagrange":
+        if self.symmetry:
+            stress_new = sym(stress_new)
 
+        if self.framework == "total-lagrange":
             # convert second to first Piola-Kirchhoff stress
             stress_new = dot(dxdX, stress_new)
 
@@ -237,10 +234,20 @@ class MaterialStrain(ConstitutiveMaterial):
         )
         dsde = np.ascontiguousarray(dsde)
 
-        if self.framework == "total-lagrange":
+        if self.symmetry:
 
+            # enforce minor symmetries on the algorithmic consistent elasticity tensor
+            dsde = np.add(dsde, np.einsum("ijlk...->ijkl...", dsde), out=dsde)
+            dsde = np.add(dsde, np.einsum("jikl...->ijkl...", dsde), out=dsde)
+            dsde = np.multiply(dsde, 0.25, out=dsde)
+
+        if self.framework == "total-lagrange":
             # convert elasticity tensor and add geometric part
             dsde = np.einsum("iI...,kK...,IJKL...->iJkL...", dxdX, dxdX, dsde)
+
+            if self.symmetry:
+                stress_new = sym(stress_new)
+
             geometric = cdya_ik(np.eye(3), stress_new)
 
             dsde = np.sum(np.broadcast_arrays(dsde, geometric), axis=0)